% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AjustEfectFix.R
\name{AjustEfectFix}
\alias{AjustEfectFix}
\title{Realizar ajustes por efectos fijos de forma interactiva}
\usage{
AjustEfectFix(df)
}
\arguments{
\item{df}{Un data.frame que contenga todas las variables potenciales para el análisis.}
}
\value{
Un data.frame (tibble) que contiene las columnas originales, los factores de
corrección generados (\code{fc_variable}) y la variable final ajustada (\code{variable_ajust}).
}
\description{
Esta función guía al usuario a través de una interfaz de consola para:
\enumerate{
\item Seleccionar variables núcleo (Y, ID, Padre, Madre).
\item Aplicar filtros numéricos opcionales.
\item Definir un modelo lineal inicial.
\item Ejecutar un proceso Stepwise (AIC) para seleccionar efectos significativos.
\item Calcular factores de corrección (FC) y generar una variable ajustada.
}
}
\details{
La función utiliza \code{step()} para la selección de variables. Si el modelo resultante
queda vacío (solo intercepto), la función se detiene con un aviso para evitar errores en
el cálculo de VIF o ANOVA. Los factores de corrección para variables continuas se basan
en la pendiente (beta), mientras que para categóricas se basan en la relación entre
medias de niveles.
}
\examples{
\dontrun{
# Supongamos que tienes un dataframe llamado 'mis_cerdos'
df_ajustado <- AjustEfectFix(mis_cerdos)
}
}
