% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGs_Mrode_Unic.R
\name{VGs_Mrode_Unic}
\alias{VGs_Mrode_Unic}
\title{Estima valores genéticos individuales en base a un registro propio único}
\usage{
VGs_Mrode_Unic(df, h2, i)
}
\arguments{
\item{df}{data.frame con registros ajustados (recomendado: correr antes AjustEfectFix()) asignados a un id}

\item{h2}{Heredabilidad}

\item{i}{Intensidad de selección}
}
\value{
Data.frame con VG, precisión y ranking
}
\description{
Esta función y todas las del paquete están basadas en el cálculo de VGs por
ayudas de selección propuestas por Mrode en su libro:
LINEAR MODELS FOR THE PREDICTION OF ANIMAL BREEDING VALUES - SECOND EDITION
}
\details{
Esta función utiliza para el cálculo del VG una medida única propia de cada animal
En la ejecución de la función se pedirá por consola la variable a evaluar
}
