% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AjustEfectFix.R
\name{AjustEfectFix}
\alias{AjustEfectFix}
\title{Realizar ajustes por los efectos fijos de un modelo lineal}
\usage{
AjustEfectFix(df)
}
\arguments{
\item{df}{Un data.frame con información de ID, Padre, Madre, Variable Y, y Variables del modelo}
}
\value{
el dataframe inicial más los factores de ajuste para los efectos utilizados y un Y_ajust
}
\description{
Esta función toma los valores reales de alguna característica y los ajusta
por los efectos significativos que indica el modelo stepwise
Además tiene un apartado para realizar filtros de variables numéricas
}
\details{
En la ejecución de la función se pedirá por consola la variable a evaluar
además también se pedirá por consola información adicional, como variables a filtrar y sus filtros
y las variables que quiere añadir al modelo stepwise
}
