% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGs_Mrode_progenie.R
\name{VGs_Mrode_progenie}
\alias{VGs_Mrode_progenie}
\title{Estima valores genéticos individuales en base a registros de progenie (para padres y madres)}
\usage{
VGs_Mrode_progenie(df, h2, i)
}
\arguments{
\item{df}{data.frame con registros ajustados (recomendado: correr antes AjustEfectFix()) asignados a un id, además información de padre y madre}

\item{h2}{Heredabilidad}

\item{i}{Intensidad de selección}
}
\value{
Data.frame con VG, precisión y ranking
}
\description{
Esta función y todas las del paquete están basadas en el cálculo de VGs por
ayudas de selección propuestas por Mrode en su libro:
LINEAR MODELS FOR THE PREDICTION OF ANIMAL BREEDING VALUES - SECOND EDITION
}
\details{
Esta función utiliza para el cálculo del VG para padres y madres, utilizando los registros de
sus diferentes crías.

En la ejecución de la función se pedirá por consola la variable a evaluar
y la columna en la cuál se encuentran los padres y madres
}
