% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGs_Mrode_repetidos.R
\name{VGs_Mrode_repetidos}
\alias{VGs_Mrode_repetidos}
\title{Estima valores genéticos individuales en base a registros propios repetidos}
\usage{
VGs_Mrode_repetidos(df, h2, repet, i)
}
\arguments{
\item{df}{data.frame con registros ajustados asignados a un id.}

\item{h2}{Heredabilidad (valor numérico entre 0 y 1).}

\item{repet}{Repetibilidad (valor numérico entre 0 y 1).}

\item{i}{Intensidad de selección (valor numérico).}
}
\value{
Data.frame con VG, precisión, progreso genético y ranking por individuo.
}
\description{
Esta función y todas las del paquete están basadas en el cálculo de VGs por
ayudas de selección propuestas por Mrode en su libro:
LINEAR MODELS FOR THE PREDICTION OF ANIMAL BREEDING VALUES - SECOND EDITION
}
\details{
Esta función utiliza para el cálculo del VG propio basado en muestras repetidas
para una variable (ej. peso al destete de diferentes camadas de una misma cerda).
}
