% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGs_Mrode_Unic.R
\name{VGs_Mrode_Unic}
\alias{VGs_Mrode_Unic}
\title{Estima valores genéticos individuales en base a un registro propio único}
\usage{
VGs_Mrode_Unic(df, h2, i)
}
\arguments{
\item{df}{data.frame con registros ajustados asignados a un id.}

\item{h2}{Heredabilidad (valor numérico entre 0 y 1).}

\item{i}{Intensidad de selección (valor numérico).}
}
\value{
Data.frame con VG, precisión, progreso genético y ranking.
}
\description{
Esta función y todas las del paquete están basadas en el cálculo de VGs por
ayudas de selección propuestas por Mrode en su libro:
LINEAR MODELS FOR THE PREDICTION OF ANIMAL BREEDING VALUES - SECOND EDITION
}
